Refactor all source code into respective packages
