+++ src/main/java/scribbles/exception/ScribblesException.java
@@ -1,4 +1,4 @@
-package ScribblesExceptions;
+package scribbles.exception;
 
 public class ScribblesException extends Exception {
     public ScribblesException(String msg) {
