+++ src/main/java/scribbles/task/EventTask.java
@@ -1,3 +1,5 @@
+package scribbles.task;
+
 import java.time.LocalDateTime;
 
 public class EventTask extends Task {
