+++ src/main/java/scribbles/task/Task.java
@@ -1,3 +1,5 @@
+package scribbles.task;
+
 import java.time.format.DateTimeFormatter;
 
 public abstract class Task {
