+++ src/main/java/scribbles/ui/Scribbles.java
@@ -1,5 +1,6 @@
+package scribbles.ui;
+
 // Utils
-import java.text.ParseException;
 import java.time.LocalDateTime;
 import java.time.format.DateTimeParseException;
 import java.util.ArrayList;
@@ -8,11 +9,16 @@ import java.util.List;
 import java.util.Scanner;
 
 // Exceptions
-import ScribblesExceptions.InvalidParamException;
-import ScribblesExceptions.MissingArgumentException;
-import ScribblesExceptions.MissingDescriptionException;
-import ScribblesExceptions.ScribblesException;
-import ScribblesExceptions.UnknownCommandException;
+import scribbles.exception.InvalidParamException;
+import scribbles.exception.MissingArgumentException;
+import scribbles.exception.MissingDescriptionException;
+import scribbles.exception.ScribblesException;
+import scribbles.exception.UnknownCommandException;
+import scribbles.storage.Storage;
+import scribbles.task.DeadlineTask;
+import scribbles.task.EventTask;
+import scribbles.task.Task;
+import scribbles.task.ToDoTask;
 
 
 public class Scribbles {
